@@ -48,12 +48,13 @@ setuptools.setup(
     },
     classifiers=[
         "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.10",
         "Operating System :: OS Independent",
     ],
     package_dir={"": "py"},
     include_package_data=True,
     packages=setuptools.find_packages(where="py"),
-    python_requires=">=3.8.0",
+    python_requires=">=3.10.0",
     entry_points={},
     install_requires=install_requires,
     extras_require=extras_require,
