@@ -1,1 +1,1 @@
-VERSION = "0.0.130"
+VERSION = "0.1.0"
